+++ src/core/layers/wfs.js
@@ -297,7 +297,7 @@ WFSLayer.prototype.filterUniqueLegendInfo = function (features, rules, legendInf
     let uniqueLegendInformation = [];
 
     rules.forEach(rule => {
-        const value = String(rule.conditions.properties[rulesKey]);
+        const value = String(rule.conditions?.properties[rulesKey]);
 
         if (conditionProperties.includes(value)) {
             const legendInformation = legendInfos.find(legendInfo => legendInfo?.label === (rule.style?.legendValue || value));