+++ CHANGELOG.md
@@ -31,7 +31,9 @@
 
 ### Fixed
 - prepareFeaturesFor3D is now called after loading
+- Geojson layer: prepareFeaturesFor3D is now called after loading.
 - Issue #1081: Transmit authorization token to services after login
+- Issue #1091: VectorStyle: read geometry type from geoserver featureTypeRequest and do not fail if some rules in style.json have no condition.
 - Issue #1093: Fixed the Wrong CRS in GeoJSON by Download filtered data
 - Issue #1099: The package.json is adapted, now the size of the bundle, created with "npm run build" is reduced.